requirements.txt
djongo==1.3.4
scrapy-djangoitem==1.1.1